# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cudium API Docs",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in to user's dashboard",
                "description": "allow user to be able to sign in returns data object with token",
                "operationId": "authsignin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Users Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "olaniyiojeyinka@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign out of user's dashboard",
                "description": "allow user to be able to sign out",
                "operationId": "authlogout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYxODk5OTQ5OCwiZXhwIjoxNjE5MDAzMDk4LCJuYmYiOjE2MTg5OTk0OTgsImp0aSI6IjRZdjlCTGdxVUhBZGpHM1kiLCJzdWIiOjgsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.yjNvT61ON8zeFwS2XUPHHXbsxYC2r6ABwbRQK3LVDH4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh user token",
                "description": "allow user to be able to refresh token",
                "operationId": "authrefresh",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYxODk5OTQ5OCwiZXhwIjoxNjE5MDAzMDk4LCJuYmYiOjE2MTg5OTk0OTgsImp0aSI6IjRZdjlCTGdxVUhBZGpHM1kiLCJzdWIiOjgsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.yjNvT61ON8zeFwS2XUPHHXbsxYC2r6ABwbRQK3LVDH4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user account",
                "description": "allow user to be able to sign up",
                "operationId": "authregister",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Users Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "User's Firstname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "User's lastname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "middlename",
                        "in": "query",
                        "description": "User's middlename",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Error occurred"
                    }
                }
            }
        },
        "/api/auth/forget": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request for token email to reset password",
                "description": "Request for token email to reset password",
                "operationId": "authforget",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Users Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Error occurred"
                    }
                }
            }
        },
        "/api/auth/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "reset password with token from email",
                "description": "Reset password with token from email",
                "operationId": "authreset",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token sent to email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "New Account Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Error occurred"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}